Strip inapplicable build args for older android versions
Prior to cordova 4.0.0, the android platform would actively reject unexpected arguments to the build. This means that just pushing all the arguments no matter the version isn't going to work if the cordova version is sufficiently old. To protect against that, we will filter the arguments passed to android builds if we detect a version of cordova < 4.0.0.