Fix build on android for cordova 3.5.0-0.2.7
The android platform pinned to version 3.5.0-0.2.7 of cordova requires that the 'android' command be on the path, and the build will fail if it is not. To work around this, if we detect that version of cordova is being used, we will append ANDROID_HOME/tools to the process's PATH (if necessary).